[FE] fix : SearchService 중복 코드 삭제

See merge request s12-bigdata-dist-sub1/S12P21A507!40

+++ back/src/main/java/com/cocoa/backend/domain/cosmetic/entity/CosmeticKeywords.java
@@ -3,7 +3,6 @@ package com.cocoa.backend.domain.cosmetic.entity;
 import java.util.Map;
 
 import com.cocoa.backend.global.converter.KeywordJsonConverter;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
@@ -16,7 +15,7 @@ import lombok.Getter;
 @Table(name = "cosmetic_keywords")
 public class CosmeticKeywords {
     @Id
-    private Long cosmeticId;
+    private Integer cosmeticId;
 
     @Column(columnDefinition = "jsonb")
     @Convert(converter = KeywordJsonConverter.class)
